+++ backend/utils/converting.go
@@ -4,10 +4,10 @@ import (
 	"strconv"
 )
 
-func StringToInt(str string) uint64 {
+func StringToInt(str string) uint {
 	intValue, err := strconv.ParseUint(str, 10, 64)
 	if err != nil {
 		return 0
 	}
-	return intValue
+	return uint(intValue)
 }
